Split integration tests into smaller SRP files
