+++ spec/integration_spec.rb
@@ -108,415 +108,6 @@ describe PerconaMigrator, integration: true do
     end
   end
 
-  context 'managing columns' do
-    let(:version) { 1 }
-
-    context 'creating column' do
-      let(:direction) { :up }
-
-      it 'adds the column in the DB table' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version
-        ).migrate
-
-        expect(:comments).to have_column('some_id_field')
-      end
-
-      it 'marks the migration as up' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version
-        ).migrate
-
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-
-    context 'droping column' do
-      let(:direction) { :down }
-
-      before do
-        ActiveRecord::Migrator.new(:up, migration_fixtures, version).migrate
-      end
-
-      it 'drops the column from the DB table' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version - 1
-        ).migrate
-
-        expect(:comments).not_to have_column('some_id_field')
-      end
-
-      it 'marks the migration as down' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version - 1
-        ).migrate
-
-        expect(ActiveRecord::Migrator.current_version).to eq(version - 1)
-      end
-    end
-
-    context 'renaming column' do
-      let(:version) { 25 }
-
-      before do
-        ActiveRecord::Migrator.run(direction, migration_path, 1)
-        ActiveRecord::Migrator.run(direction, migration_path, 2)
-      end
-
-      it 'changes the column name' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).to have_column('new_id_field')
-      end
-
-      it 'does not keep the old column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).not_to have_column('some_id_field')
-      end
-    end
-  end
-
-  context 'creating references' do
-    context 'when no option is set' do
-      let(:version) { 16 }
-
-      it 'adds a reference column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).to have_column('user_id')
-      end
-    end
-
-    context 'when polymorphic is set to true' do
-      let(:version) { 17 }
-
-      it 'adds a column for the id' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).to have_column('user_id')
-      end
-
-      it 'adds a column for the type' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).to have_column('user_type')
-      end
-
-      context 'and index is set to true' do
-        let(:version) { 19 }
-
-        it 'adds a compound index for both the id and type columns' do
-          ActiveRecord::Migrator.run(direction, migration_path, version)
-
-          expect(:comments)
-            .to have_index('index_comments_on_user_id_and_user_type')
-        end
-      end
-    end
-
-    context 'when index is set to true' do
-      let(:version) { 18 }
-
-      it 'adds an index for the reference column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-
-        expect(:comments).to have_index('index_comments_on_user_id')
-      end
-    end
-  end
-
-  context 'removing references' do
-    let(:version) { 20 }
-
-    before { ActiveRecord::Migrator.run(direction, migration_path, 16) }
-
-    context 'when no option is set' do
-      it 'removes the reference column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).not_to have_column('user_id')
-      end
-    end
-
-    context 'when polymorphic is set to true' do
-      it 'removes the reference id column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).not_to have_column('user_id')
-      end
-
-      it 'removes the reference type column' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).not_to have_column('user_type')
-      end
-    end
-  end
-
-  context 'managing indexes' do
-    let(:version) { 2 }
-
-    context 'adding indexes' do
-      let(:direction) { :up }
-
-      # TODO: Create it directly like this?
-      before do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          1
-        ).migrate
-      end
-
-      it 'executes the percona command' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version
-        ).migrate
-
-        expect(:comments).to have_index('index_comments_on_some_id_field')
-      end
-
-      it 'marks the migration as up' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version
-        ).migrate
-
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-
-    context 'removing indexes' do
-      let(:direction) { :down }
-
-      before do
-        ActiveRecord::Migrator.new(
-          :up,
-          migration_fixtures,
-          1
-        ).migrate
-
-        ActiveRecord::Migrator.new(
-          :up,
-          migration_fixtures,
-          version
-        ).migrate
-      end
-
-      it 'executes the percona command' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version - 1
-        ).migrate
-
-        expect(:comments).not_to have_index('index_comments_on_some_id_field')
-      end
-
-      it 'marks the migration as down' do
-        ActiveRecord::Migrator.new(
-          direction,
-          migration_fixtures,
-          version - 1
-        ).migrate
-
-        expect(ActiveRecord::Migrator.current_version).to eq(1)
-      end
-    end
-
-    context 'renaming indexes' do
-      let(:direction) { :up }
-      let(:version) { 13 }
-
-      before do
-        ActiveRecord::Migrator.new(:up, migration_fixtures, 2).migrate
-      end
-
-      it 'executes the percona command' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(:comments).to have_index('new_index_comments_on_some_id_field')
-      end
-
-      it 'marks the migration as down' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-  end
-
-  context 'adding/removing unique indexes' do
-    let(:version) { 3 }
-
-    context 'adding indexes' do
-      let(:direction) { :up }
-
-      before do
-        ActiveRecord::Migrator.new(:up, migration_fixtures, 1).migrate
-      end
-
-      it 'executes the percona command' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-
-        expect(unique_indexes_from(:comments))
-          .to match_array(['index_comments_on_some_id_field'])
-      end
-
-      it 'marks the migration as up' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-
-    context 'removing indexes' do
-      let(:direction) { :down }
-
-      before do
-        ActiveRecord::Migrator.run(:up, migration_path, 1)
-        ActiveRecord::Migrator.run(:up, migration_path, version)
-      end
-
-      it 'executes the percona command' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-
-        expect(unique_indexes_from(:comments))
-          .not_to match_array(['index_comments_on_some_id_field'])
-      end
-
-      it 'marks the migration as down' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(ActiveRecord::Migrator.current_version).to eq(1)
-      end
-    end
-  end
-
-  context 'creating a table' do
-    let(:version) { 8 }
-
-    it 'creates the table' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(tables).to include('things')
-    end
-
-    it 'marks the migration as up' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(ActiveRecord::Migrator.current_version).to eq(version)
-    end
-  end
-
-  context 'dropping a table' do
-    let(:version) { 8 }
-    let(:direction) { :down }
-
-    it 'drops the table' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(tables).not_to include('things')
-    end
-
-    it 'updates the schema_migrations' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(ActiveRecord::Migrator.current_version).to eq(0)
-    end
-  end
-
-  context 'when changing column null' do
-    let(:direction) { :up }
-    let(:column) do
-      columns(:comments).find { |column| column.name == 'some_id_field' }
-    end
-
-    before do
-      ActiveRecord::Migrator.new(:up, migration_fixtures, 1).migrate
-    end
-
-    context 'when null is true' do
-      let(:version) { 14 }
-
-      it 'sets the column to allow nulls' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(column.null).to be_truthy
-      end
-
-      it 'marks the migration as up' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-
-    context 'when null is false' do
-      let(:version) { 15 }
-
-      it 'sets the column not to allow nulls' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(column.null).to be_falsey
-      end
-
-      it 'marks the migration as up' do
-        ActiveRecord::Migrator.run(direction, migration_path, version)
-        expect(ActiveRecord::Migrator.current_version).to eq(version)
-      end
-    end
-  end
-
-  context 'adding timestamps' do
-    let(:version) { 22 }
-
-    it 'adds a created_at column' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(:comments).to have_column('created_at')
-    end
-
-    it 'adds a updated_at column' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(:comments).to have_column('updated_at')
-    end
-  end
-
-  context 'removing timestamps' do
-    let(:version) { 23 }
-
-    before do
-      ActiveRecord::Migrator.run(direction, migration_path, 22)
-    end
-
-    it 'removes the created_at column' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(:comments).not_to have_column('created_at')
-    end
-
-    it 'removes the updated_at column' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(:comments).not_to have_column('updated_at')
-    end
-  end
-
-  context 'renaming a table' do
-    let(:version) { 24 }
-
-    before do
-      ActiveRecord::Migrator.run(direction, migration_path, 1)
-      ActiveRecord::Migrator.run(direction, migration_path, 2)
-    end
-
-    it 'changes the table name' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(tables).to include('new_comments')
-    end
-
-    it 'does not keep the old name' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(tables).not_to include('comments')
-    end
-
-    it 'changes the index names in the new table' do
-      ActiveRecord::Migrator.run(direction, migration_path, version)
-      expect(:new_comments).to have_index('index_new_comments_on_some_id_field')
-    end
-  end
-
   context 'when the migration failed' do
     context 'and the migration is not an alter table statement' do
       let(:version) { 8 }
